Update README.md to accomodate subdirectory installation behaviour
Edited paragraph to accommodate the change in template name if user declines default and specifies a subdirectory location for the installation.

I'll leave a question mark hanging in the air about the .docx and .html children of the quarto template. Their names are not modified on installation, which might be on purpose, but they could be added to .gitignore if not?